# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Live Score — Housing Trust Infrastructure API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Authentication required
        "404":
          description: User not found
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid profile input
        "401":
          description: Authentication required
        "404":
          description: User not found
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Authentication required
        "403":
          description: Administrator access required
...
  /users/me/verify/phone:
    post:
      operationId: verifyPhone
      tags: [users]
      summary: Reject direct phone verification; use the OTP flow
      security:
        - BearerAuth: []
      responses:
        "401":
          description: Authentication required
        "403":
          description: Phone verification is available only through OTP
        "404":
          description: User not found
        "409":
          description: Phone is already verified
  /users/me/verify/identity:
    post:
      operationId: verifyIdentity
      tags: [users]
      summary: Submit an identity document for administrator review
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationSubmission"
      responses:
        "202":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "400":
          description: Invalid or unsupported document
        "401":
          description: Authentication required
        "403":
          description: Document belongs to another user
        "409":
          description: Verification is already pending or complete
        "503":
          description: Private object storage is unavailable
  /users/me/verify/income:
    post:
      operationId: verifyIncome
      tags: [users]
      summary: Submit an income document for administrator review
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationSubmission"
      responses:
        "202":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "400":
          description: Invalid or unsupported document
        "401":
          description: Authentication required
        "403":
          description: Document belongs to another user
        "409":
          description: Verification is already pending or complete
        "503":
          description: Private object storage is unavailable
  /users/me/verifications:
    get:
      operationId: listMyVerifications
      tags: [users]
      summary: List the current user's verification applications
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Verification applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VerificationApplication"
        "401":
          description: Authentication required
  /users/me/verifications/{applicationId}/document:
    get:
      operationId: getMyVerificationDocument
      tags: [users]
      summary: Stream a private verification document for its owner
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: applicationId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Private document
          content:
            application/pdf: {}
            image/jpeg: {}
            image/png: {}
            image/webp: {}
        "401":
          description: Authentication required
        "404":
          description: Document not found
        "503":
          description: Private document storage is unavailable
  /users/{userId}/verify/{verificationType}/decision:
    post:
      operationId: decideVerification
      tags: [users]
      summary: Approve or reject a pending verification application
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
        - in: path
          name: verificationType
          required: true
          schema:
            type: string
            enum: [identity, income]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationDecisionInput"
      responses:
        "200":
          description: Decision result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "400":
          description: Invalid decision or missing rejection comment
        "401":
          description: Authentication required
        "403":
          description: Administrator access required
        "404":
          description: User not found
        "409":
          description: Application was already decided
  /users/verifications:
    get:
      operationId: listVerificationQueue
      tags: [users]
      summary: List verification applications for an administrator
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected, all]
            default: pending
      responses:
        "200":
          description: Verification queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VerificationQueueItem"
        "401":
          description: Authentication required
        "403":
          description: Administrator access required
  /users/verifications/{applicationId}/document:
    get:
      operationId: getVerificationDocument
      tags: [users]
      summary: Stream a verification document for an administrator
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: applicationId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Private document
          content:
            application/pdf: {}
            image/jpeg: {}
            image/png: {}
            image/webp: {}
        "401":
          description: Authentication required
        "403":
          description: Administrator access required
        "404":
          description: Document not found
        "503":
          description: Private document storage is unavailable
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, phone, name, pointsBalance, status, liveScore, verificationLevel, isAdmin, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        pointsBalance:
          type: integer
        bonusBalanceRub:
          type: number
        status:
          type: string
          enum: [novice, silver, gold, platinum]
        statusMultiplier:
          type: number
        pointsToNextStatus:
          type: ["integer", "null"]
        nextStatus:
          type: ["string", "null"]
        liveScore:
          type: integer
        verificationLevel:
          type: integer
        isPhoneVerified:
          type: boolean
        isIdentityVerified:
          type: boolean
        isIncomeVerified:
          type: boolean
        identityVerificationStatus:
          $ref: "#/components/schemas/VerificationStatus"
        incomeVerificationStatus:
          $ref: "#/components/schemas/VerificationStatus"
        isAdmin:
          type: boolean
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    VerificationStatus:
      type: string
      enum: [not_started, pending, approved, rejected]
...
    VerificationSubmission:
      type: object
      required: [objectPath, fileName, contentType, fileSize]
      properties:
        objectPath:
          type: string
        fileName:
          type: string
          maxLength: 160
        contentType:
          type: string
          enum: [application/pdf, image/jpeg, image/png, image/webp]
        fileSize:
          type: integer
          minimum: 1
          maximum: 10485760
    VerificationApplication:
      type: object
      required: [id, verificationType, status, fileName, contentType, fileSize, createdAt]
      properties:
        id:
          type: integer
        verificationType:
          type: string
          enum: [identity, income]
        status:
          $ref: "#/components/schemas/VerificationStatus"
        fileName:
          type: string
        contentType:
          type: string
        fileSize:
          type: integer
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
        decidedAt:
          type: ["string", "null"]
    VerificationQueueItem:
      allOf:
        - $ref: "#/components/schemas/VerificationApplication"
        - type: object
          required: [user]
          properties:
            user:
              type: ["object", "null"]
              properties:
                id:
                  type: integer
                name:
                  type: string
                phone:
                  type: string
    VerificationDecisionInput:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean
        comment:
          type: string
          maxLength: 1000
...
    VerifyResponse:
      type: object
      required: [success, verificationType, status, scoreChange, newScore, applicationId]
      properties:
        success:
          type: boolean
        verificationType:
          type: string
          enum: [identity, income]
        status:
          $ref: "#/components/schemas/VerificationStatus"
        scoreChange:
          type: integer
        newScore:
          type: integer
        applicationId:
          type: integer
        application:
          $ref: "#/components/schemas/VerificationApplication"
        alreadyDone:
          type: boolean
...