# lib/api-zod/src/generated/types/verificationStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Live Score — Housing Trust Infrastructure API
 * OpenAPI spec version: 0.2.0
 */

export type VerificationStatus = typeof VerificationStatus[keyof typeof VerificationStatus];


export const VerificationStatus = {
  not_started: 'not_started',
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;